modularizing (of sort) timemix
